// frontend/src/types/types.ts
export interface Patient {
  /**
   * Id
   * @format uuid
   */
  id: string;
  /**
   * Hospital
   * @format uuid
   */
  hospital: string;
  /**
   * First name
   * @minLength 1
   * @maxLength 100
   */
  first_name: string;
  /**
   * Last name
   * @minLength 1
   * @maxLength 100
   */
  last_name: string;
  /**
   * Dob
   * @format date
   */
  dob: string;
  /** Sex */
  sex: "M" | "F" | "O";
  /**
   * Phone
   * @maxLength 20
   */
  phone?: string | null;
  /**
   * Email
   * @format email
   * @maxLength 254
   */
  email?: string | null;
}